Move graphviz to better location
