+++ src/easydmp/dmpt/models.py
@@ -25,8 +25,6 @@ from django.utils.safestring import mark_safe
 from django.utils.html import format_html, escape
 from django.utils.text import slugify
 
-from flow.graphviz import _prep_dotsource, view_dotsource, render_dotsource_to_file
-
 from .errors import TemplateDesignError
 from .utils import DeletionMixin
 from .utils import RenumberMixin
@@ -34,6 +32,7 @@ from .utils import print_url
 from .utils import render_from_string
 
 from easydmp.eestore.models import EEStoreCache
+from easydmp.lib.graphviz import _prep_dotsource, view_dotsource, render_dotsource_to_file
 from easydmp.lib.models import ModifiedTimestampModel
 from easydmp.utils import pprint_list
